refactor: colocate web ui tests
