+++ web/src/app/shell/__tests__/Sidebar.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi } from "vitest";
 import { render, screen, fireEvent, cleanup } from "@testing-library/react";
-import { Sidebar } from "../app/shell/Sidebar";
+import { Sidebar } from "../Sidebar";
 
 describe("Sidebar", () => {
   it("sidebar_renders_brand_new_session_and_runtime_status", () => {
